requirements.txt
3==0.0.0
aiohttp==3.7.4.post0
asgiref==3.5.2
async-timeout==3.0.1
attrs==22.1.0
boto==2.49.0
boto3==1.26.48
botocore==1.29.48
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
discord==1.7.3
discord.py==1.7.3
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.1
django-extensions==3.2.1
django-filter==22.1
django-jazzmin==2.6.0
django-storages==1.13.2
filelock==3.8.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
multidict==6.0.2
platformdirs==2.5.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
virtualenv==20.16.5
waitress==2.1.2
whitenoise==6.3.0
yarl==1.8.1
